api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-magic==0.4.27
ffmpeg-python==0.2.0
Pillow==10.2.0
mutagen==1.47.0
pydantic==2.5.3
aiofiles==23.2.1
redis==5.0.1
prometheus-client==0.19.0
psutil==5.9.6
httpx==0.26.0
pytest==7.4.3
pytest-asyncio==0.21.1
